chore: refactor utils
